fix: updated test cases
